Add files via upload
I have made it so the program can tell the user weather the solution can be found with 2 numbers or not. This assumes that - and / will not be needed and the whole method seems very inefficient. I would like to put the operators + - * / into an array and call them in a for loop but I cannot find seem to find anything useful online and cannot work it out.